programacao/arts/pyproject.toml
[tool.poetry]
name = "artes"
version = "0.1.0"
description = ""
authors = ["Arthur Alvim <afmalvim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
